added user registeration with JWT

+++ resources/item.py
@@ -2,8 +2,11 @@ import uuid
 from flask import Flask, request
 from flask_smorest import abort, Blueprint
 from flask.views import MethodView
+from flask_jwt_extended import jwt_required, get_jwt
+
 from schemas import ItemSchema, ItemUpdateSchema
 from models.item import ItemModel
+
 from db import db
 from sqlalchemy.exc import SQLAlchemyError, IntegrityError
 
@@ -16,7 +19,12 @@ class Item(MethodView):
         item = ItemModel.query.get_or_404(item_id)
         return item
 
+    @jwt_required(fresh=True)
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
+
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -44,6 +52,7 @@ class ItemList(MethodView):
     def get(self):
         return ItemModel.query.all()
 
+    @jwt_required(fresh=True)
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
